requirements.txt
flask
flask_socketio
yfinance
twilio
schedule
python-dotenv
imaplib2
pandas
matplotlib
requests
nginx
